+++ src/bridge/MatrixHandler.ts
@@ -1039,7 +1039,7 @@ export class MatrixHandler {
 
         let cacheBody = ircAction.text;
 
-        // special handling for replies
+        // special handling for replies (and threads)
         if (event.content["m.relates_to"] && event.content["m.relates_to"]["m.in_reply_to"]) {
             const eventId = event.content["m.relates_to"]["m.in_reply_to"].event_id;
             const reply = await this.textForReplyEvent(event, eventId, ircRoom);
@@ -1261,20 +1261,25 @@ export class MatrixHandler {
 
     private async textForReplyEvent(event: MatrixMessageEvent, replyEventId: string, ircRoom: IrcRoom):
     Promise<{formatted: string; reply: string}|null> {
-        const REPLY_REGEX = /> <(.*?)>(.*?)\n\n([\s\S]*)/;
+        // strips out the quotation of the original message, if needed
+        const replyText = (body: string): string => {
+            const REPLY_REGEX = /> <(.*?)>(.*?)\n\n([\s\S]*)/;
+            const match = REPLY_REGEX.exec(body);
+            if (match === null || match.length !== 4) {
+                return body;
+            }
+            return match[3];
+        };
+
         const REPLY_NAME_MAX_LENGTH = 12;
         const eventId = replyEventId;
         if (!event.content.body) {
             return null;
         }
-        const match = REPLY_REGEX.exec(event.content.body);
-        if (match === null || match.length !== 4) {
-            return null;
-        }
 
+        const rplText = replyText(event.content.body);
         let rplName: string;
         let rplSource: string;
-        const rplText = match[3];
         let cachedEvent = this.getCachedEvent(eventId);
         if (!cachedEvent) {
             // Fallback to fetching from the homeserver.
@@ -1289,8 +1294,7 @@ export class MatrixHandler {
                 const isReply = eventContent.content["m.relates_to"] &&
                     eventContent.content["m.relates_to"]["m.in_reply_to"];
                 if (isReply) {
-                    const sourceMatch = REPLY_REGEX.exec(eventContent.content.body);
-                    rplSource = sourceMatch && sourceMatch.length === 4 ? sourceMatch[3] : event.content.body;
+                    rplSource = replyText(eventContent.content.body);
                 }
                 else {
                     rplSource = eventContent.content.body;