+++ widget/src/ProvisioningApp.tsx
@@ -1,6 +1,5 @@
 import React, { createContext, useEffect, useMemo, useState } from 'react';
 import { WidgetApi, WidgetApiToWidgetAction } from 'matrix-widget-api';
-import urlJoin from 'url-join';
 
 import { ProvisioningClient, ProvisioningError } from './ProvisioningClient';
 import * as Text from './components/text';
@@ -73,7 +72,7 @@ export const ProvisioningApp: React.FC<React.PropsWithChildren<{
             return;
         }
 
-        const widgetApi = new WidgetApi(widgetId);
+        const widgetApi = new WidgetApi(widgetId, '*');
         widgetApi.on('ready', () => {
             console.log('Widget API ready');
         });